Properly clamp SPC CPU registers and temporaries.
Improper bounds checking on some emulated operations of the SPC700 audio
CPU can lead to arbitary code execution if malicious SPC audio is fed to
the emulator.

For more details see
https://scarybeastsecurity.blogspot.gr/2016/12/redux-compromising-linux-using-snes.html

More work will be needed to audit the rest of libgme but this at least
should close the known point(s) of vulnerability.

Thanks to scarybeast for the patch (and for spotting the flaws)!