spc: Use exact hardware register widths.
Based on input from scarybeasts (who discovered the SPC emulator
security flaw fixed for 0.6.1).  The basic idea is to use types for the
emulated registers and memory I/O that are exactly as wide as what was
actually present on the S-SMP hardware itself.

While there were performance considerations that led to using 'int' and
'unsigned' in the emulator, the revised code is not noticeably slower
with modern CPUs/compilers, in some situations it's actually faster.