Avoid undefined behavior with left-shift of negative ints.
Caught by UBSan when testing the library using the provided gme_player
sample app.

While this has not caused improper code to be generated, it's still better to
stay ahead of the compiler writers...

Interestingly, part of this was done back in 2013 (commit
6e1544d79a559cee6ff6b98efcc3df41ff27df66) in response to a PVS Studio
check I'd happened to run across, see
http://www.viva64.com/en/examples/V610/.

But I'd missed this one back then when grepping for it, probably due to
the different variable name. :(