multichannel: Add simple test suite.
Add a simple test suite intended to help iron out whether an upgraded
libgme.so causes differences in audio output.

So far in the .spcs and .gyms I've tested the multichannel changes do
not appear to break existing code (note this is testing with the
emulator not in multichannel mode, but still using the new code).

There's tons more that could be done here (e.g. checking for performance
regressions, automatically comparing git revisions between each other)
but this is more than what we had 6 hours ago...

This is pretty specific to a Unix-style toolchain so I haven't tied it
into CMake.  But I've tried to make it easy to step through at least.