nsfe: Add some additional hardening and sanity checks.
It's not safe to rely on debugging assertions alone in face of hostile
user input so try to add additional sanity checks in NSFE emulator.

I have no specific exploits for these to defend against, and there are
probably better ways to achieve this same effect, but it seems prudent
to take additional precautions given the nature of the flaw fixed in the
previous commit.