zlib: Minor cleanups.
Some of the changes I want are just easier to handle myself, especially
since they are very minor or just things that were already 'bad' in the
existing code.  But since we're changing the code anyways, now is a good
time to make the improvement.

- Be more careful with calls to read/write functions that we don't
accidentally cast a (long)-1 to (size_t) (which would wrap to a very
large value).
- prefer C++ style casts
- mark some added variables for ABI (adding members to existing classes
would force me to bump the SOVERSION since it's a binary incompatible
change, which I would rather not do yet).  For a similar reason I'll
actually have to keep the GZip_* classes for now, otherwise a libgme
upgrade could break application code that had already been linked to it.