Fix build failure in C-based demo player.
The change to make BLARGG_EXPORT apply to specific symbols breaks
compilation for C-based uses of gme.h. This might also be related to
newer compilers being stricter about the overloading of min/max on
multiple types (which happens in blargg_source.h and is not permitted in
C). Problem discovered by the build pipeline on Bitbucket.

I fix this by removing min/max from blargg_source.h (so it can be
included from C or C++ code) and using the C++ stdlib instead of making
a custom reimplementation of min/max.