WIP: Add a function to avoid automatic track ending.
This maintains automatic track ending as the default to address issue #25
while making it easier to restore the previous default behavior of
having music loop infinitely.

To maintain binary compatibility on Music_Emu and its subclasses, I make
this a global library setting that can be queried and set (or reset),
instead of specific to a Music_Emu class.

This should resolve issue #29.