Add a function to avoid automatic track ending.
This maintains automatic track ending as the default to address issue #25
while making it easier to restore the previous default behavior of
having music loop infinitely.

This changes the structure of the Music_Emu object itself, but that has
not been part of the binary interface (the C API from gme.h treats
Music_Emu as an opaque handle). So as long as you don't mix two versions
of the same library in the same application somehow then this shouldn't
cause existing compiled apps to break.

Fixes issue #29. Thanks to Wohlstand for reviewing the patch!