resampler: Don't resample when sample rates are matched.
The FIR resampler is always in use on some emulators (VGM in particular
but also SPC). The resampler produces high quality output when we have
to resample, but we don't always have to resample and in that situation
it's better to avoid adding noise.

This is a rework of a patch Wohlstand submitted [1] since it felt
cleaner to me and doesn't rely on floating point arithmetic to be exact.

This fixes #34. I can't even tell apart the "bad" and "good" quality
samples though so if I've still got it wrong, please reopen.

[1] https://bitbucket.org/Wohlstand/game-music-emu/commits/03d70998bbbb065231c8c70f58c62ccd24e6896a