example: Use SDL2 rendering to draw audio scope and simplify.
I think a lot of this code was copied from a more generic oscilloscope
library when this was first imported as it's quite generic. I simplified
the code a bunch to cover how it's actually used.

From there I let SDL's render subsystem take care of drawing the lines
properly rather than trying to understand the existing hand-optimized
code (which was tied intimately to paletted pixel formats anyways).