Parse hours and milliseconds
Added a parser for milliseconds, which are written as fractions of a second, as many of the m3u files I've come across have their times formatted as h:mm:ss.fff.
Also changed the if statement for detecting a colon into a while statement, since there are usually 2 colons, one for hours and one for minutes.

Because time is handled down to the millisecond in the m3u parsing phase, I'm skipping the step in Gme_file.cpp where various time lengths are multiplied by 1000.

Also edited test.m3u to reflect these time formatting conventions.