Removed track fade from test.m3u
While the parser collects track fade info successfully, the player demo uses the default fade time, so any fade time specified by the user will be ignored.
The player could be altered to use the m3u's fade duration, but that seems to be outside the scope of this branch for the time being.