test: Make the demo app into a simple testcase whose output is actually checked.
This should test building the library, linking the library to a sample
app, using the library (and its NSF emulator), and ensuring that the NSF
output doesn't change from the output that would have been produced by
the library before I introduced my security hardening fixes (though,
only on NSF which was barely affected by these changes).

But now that there is a base for testing, we can add known inputs (and
their checksum) and their corresponding known outputs (and their
checksum) to the test suite as time goes on, and provide coverage for
the other formats.

If rendering should improve, we can still change the checksums at that
time.