RSN: Check for whether track info is in bounds.
Although the existing code should be correct (unless the application
itself is trying to ask for missing tracks, which is outside our threat
model), I still feel better with an explicit check.

The blargg_vector change is just because I happened to be in there
checking on its size() function.