Effects_Buffer: Use sample offsets for multi-channel buffering
If we request more than 1 frame of multi-channel audio, the buffer
will use a method that involves buffering 2 samples at a time for each
voice, The placement of these samples, however, is incorrect.

Each frame should consist of 1 sample per voice (we actually
duplicate the sample for the sake of stereo upmixing, so if
`max_voices` equals 8, the frame will consist of 16 samples in total).

Rather than placing the voice's samples right next to each other,
we should instead give the sample(s) for the 2nd frame an offset of
`max_voices` (times 2 for upmixing), in order to account for the rest
of the voices that the frame will consist of. If we don't do this,
the two frames will become interlaced.

Reasons why this has gone unnoticed:
- Multi-channel audio is a niche feature.
- In most cases, only 1 frame per request is needed.
- Even with every 2 frames interlaced, it's not immediately obvious that
  anything is wrong with the audio, particularly 8-bit audio, unless
  it is either closely inspected or upsampled.